+++ CHANGELOG.md
@@ -53,6 +53,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - x11: Set WM_CLASS property ([#1868] by [@psychon])
 - Expose `RawWindowHandle` for `WindowHandle` under the `raw-win-handle` feature ([#1828] by [@djeedai])
 - `Slider` widget now warns if max < min and swaps the values ([#1882] by [@Maan2003])
+- Widget/Slider: Add stepping functionality ([#1875] by [@raymanfx])
 
 ### Changed